requirements.txt
avro-python3
Django<3
djangorestframework
protobuf
requests
